# perf: Cache requirements for pods alongside requests